pyproject.toml
[project]
name = "sovd-gateway-tests"
version = "0.0.1"
description = "Integration tests for sovd-gateway"
authors = [{ name = "SOVD Contributors" }]
readme = "tests/README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.12"
dependencies = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "requests>=2.25.0",
    "httpx>=0.27.0",
    "playwright>=1.40.0",
    "pytest-playwright-asyncio>=0.1.0",
    "pytest-httpserver>=1.1.3",
    "cryptography>=45.0.6",
    "pyjwt>=2.10.1",
]

[tool.setuptools]
py-modules = []

[tool.pytest.ini_options]
minversion = "8.0"
asyncio_mode = "auto"
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
    "--tb=short",
]
testpaths = ["tests"]
markers = [
    "requires_binary: marks tests that require the sovd-gateway binary",
    "ui: marks tests as UI tests using playwright",
]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
    "ignore::ResourceWarning",
]

[tool.coverage.run]
branch = true
source = ["tests"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "N",   # pep8-naming
    "S",   # flake8-bandit
    "BLE", # flake8-blind-except
    "FBT", # flake8-boolean-trap
    "A",   # flake8-builtins
    "COM", # flake8-commas
    "DTZ", # flake8-datetimez
    "EM",  # flake8-errmsg
    "ISC", # flake8-implicit-str-concat
    "PIE", # flake8-pie
    "T20", # flake8-print
    "Q",   # flake8-quotes
    "RSE", # flake8-raise
    "RET", # flake8-return
    "SLF", # flake8-self
    "SIM", # flake8-simplify
    "TID", # flake8-tidy-imports
    "PTH", # flake8-use-pathlib
]
ignore = [
    "T20",    # Allow print statements in tests
    "S603",   # Allow subprocess
    "S101",   # Allow assert statements
    "FBT001", # Boolean positional arg in function definition
    "FBT002", # Boolean default positional argument in function definition
    "COM812", # Trailing comma
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.uv]
default-groups = ["tools"]

[dependency-groups]
tools = [
    "pre-commit>=4.2",
    "ruff>=0.12.9",
    "ty>=0.0.1a26",
]
